Still workign through datatypes...
